+++ integration-tests/pom.xml
@@ -304,6 +304,7 @@
                    **/ItIstioDomainInPV,
                    **/ItIstioMiiDomain,
                    **/ItMiiAuxiliaryImage,
+                   **/ItMiiCreateAuxImageWithImageTool,
                    **/ItMiiAuxiliaryImageCluster,
                    **/ItMiiSampleWlsAux,
                    **/ItMiiServiceMigration,
@@ -329,6 +330,7 @@
                   **/ItCrossDomainTransaction,
                   **/ItKubernetesEvents,
                   **/ItMiiAuxiliaryImage,
+		  **/ItMiiCreateAuxImageWithImageTool,
                   **/ItMiiAuxiliaryImageCluster,
                   **/ItMiiDomain,
                   **/ItMiiDomainModelInPV,
@@ -399,6 +401,7 @@
                   **/ItManagedCoherence,
                   **/ItManageNameSpace,
                   **/ItMiiAuxiliaryImage,
+                  **/ItMiiCreateAuxImageWithImageTool,
                   **/ItMiiAuxiliaryImageCluster,
                   **/ItMiiCustomSslStore,
                   **/ItMiiDomain,
@@ -483,8 +486,8 @@
                 <includes-failsafe>
                     **/ItMultiDomainModels,
                     **/ItMiiUpdateDomainConfig,
-                    **/ItMiiAuxiliaryImage40,
-                    **/ItMiiNewCreateAuxImage40,
+                    **/ItMiiAuxiliaryImage,
+                    **/ItMiiCreateAuxImageWithImageTool,
                     **/ItMiiDynamicUpdate*,
                     **/ItMiiSampleWlsMain,
                     **/ItMiiSampleWlsAux,
